+++ src/eglib/hal/four_wire_spi/libopencm3_stm32f4.c
@@ -251,7 +251,7 @@ static void send(
 	eglib_t *eglib,
 	hal_dc_t dc,
 	uint8_t *bytes,
-	uint16_t length
+	uint32_t length
 ) {
 	four_wire_spi_libopencm3_stm32f4_config_t *config;
