+++ src/eglib/hal/i2c/libopencm3_stm32f4.c
@@ -215,7 +215,7 @@ static void send(
 	eglib_t *eglib,
 	hal_dc_t dc,
 	uint8_t *bytes,
-	uint16_t length
+	uint32_t length
 ) {
 	if(hal_i2c_send_slave_addr(eglib))
 		send_slave_address(eglib, dc);
